Basic space scene and ship controller.
Added a basic tiling space background with four walls, and a camera that follows the spaceship. The spaceship is controlled by moving forward and backward with W and S. It is turned left and right with A and D. You can shoot with LMB. You can also set the max speed, max rotational speed, acceleration speed, rotational acceleration speed, and you can toggle uniform rotation, uniform movement and whether the bullet aligns with the ship or points at the cursor.