Revised the spaceship movement to allow for the ship's rotation to align to your inputs.
The controlled rotation boolean in the spaceship's metadata now allows you to keep the old system, or have the ship move in the direction of your WASD input, and then it rotates accordingly to line up with that.

The default is also currently set to uniform rotation, but non-uniform movement, so the ship does use acceleration (this can be adjusted though). I have also added in a new drag variable that reduces your spaceship's speed by the given amount each second on whichever axes currently have no directional input.