Added wormholes with travel popup
I added in a very basic wormhole with a UI popup in the bottom of the screen asking if you want to travel through when you are close enough to the wormhole.

I also added a boost option for the spaceship to move faster while holding in shift, and I set it so that by default, you move in the direction of your WASD input, buy your spaceships looks and shoots at your cursor.